v4/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
pandas
spotipy
openpyxl